# Conflicts:
#	fearless/Modules/Staking/StakingBondMore/View/StakingBondMoreViewLayout.swift
#	fearless/Modules/Staking/StakingBondMoreConfirmation/Flow/Parachain/StakingBondMoreConfirmationParachainViewModelState.swift
#	fearless/Modules/Staking/StakingUnbondConfirm/Flow/Parachain/StakingUnbondConfirmParachainViewModelFactory.swift
#	fearless/Modules/Staking/StakingUnbondSetup/View/StakingUnbondSetupLayout.swift
#	fearless/en.lproj/Localizable.strings

+++ fearless/Modules/AccountImport/Model/AccountImportJsonFactory.swift
@@ -20,7 +20,7 @@ final class AccountImportJsonFactory {
             networkTypeConfirmed = true
         } else {
             if let chainType = info.chainType {
-                chain = SNAddressType(rawValue: UInt8(chainType))?.chain
+                chain = SNAddressType(rawValue: UInt16(chainType))?.chain
             } else {
                 chain = nil
             }
